[refactor] ThemePrefContext and DataContext moved to context folder
