+++ Framework/Sample/SampleMenusPage.js
@@ -11,7 +11,7 @@ import Picker from '../UI/Picker';
 import DropdownMenu from '../UI/DropdownMenu';
 import DropdownCheckMenu from '../UI/DropdownCheckMenu';
 // data
-import { DataContext } from '../Common/DataContext';
+import { DataContext } from '../Contexts/DataContext';
 // const
 const PICKER_ITEM_LIST = [
   { label: 'Red', value: 'red' },
